benchmark/analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.4.1",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "polars-lts-cpu>=1.9.0",
]

[tool.ruff]
line-length = 80

[tool.uv]
dev-dependencies = [
    "ipython>=8.27.0",
]
